Refactor merge_sort.cpp and LinkedList class
This commit refactors the merge_sort.cpp file and the LinkedList class.

In merge_sort.cpp, the code has been modified to include the necessary headers for vector and to remove unnecessary comments. The merge() function has been removed, as it is not being used.

In the LinkedList class, a new struct Node has been defined to represent a node in the linked list. The class now has a head pointer to the first node, and a constructor to initialize an empty list. The insert_last() function has been implemented to insert a new node at the end of the list. The printList() function has been updated to correctly print the elements of the list. The getMiddle() function has been added to find the middle node of the list using the fast and slow pointer pattern. The mergeSort() function has been added, but it is currently empty.

These changes improve the code structure and functionality of the merge_sort.cpp file and the LinkedList class.